# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Kaftar HTTP API
  version: 1.0.0
  description: |
    Human-maintained contract for the migrated service. A successful send means
    committed to PostgreSQL, not delivered or necessarily published to RabbitMQ. All routes are unauthenticated; use a trusted network.
    Status routes look up IDs globally, without filtering by their route channel.
servers:
  - url: http://localhost:1404
    description: Local Docker Compose
security: []
paths:
...
  /api/v1/http/send:
    post:
      operationId: sendHTTP
      summary: Queue an outbound HTTP request
      description: |
        Request bodies are sent as literal strings. This provider is disabled by
        default. Delivery requires an operator-allowlisted host/port and, for private
        destinations, a CIDR exception. Redirects and cookie persistence are disabled.
        Policy failures appear in delivery status; do not expose the API to untrusted callers.
      parameters:
        - $ref: '#/components/parameters/MaxRetries'
        - $ref: '#/components/parameters/SendAt'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HTTPMessage'
      responses:
        '200':
          $ref: '#/components/responses/Queued'
        '409':
          $ref: '#/components/responses/Conflict'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/SendFailure'
        '503':
          $ref: '#/components/responses/Unavailable'
...
  /api/v1/http/status:
    get:
      operationId: statusHTTP
      summary: Get message status by ID
      parameters:
        - $ref: '#/components/parameters/MessageID'
      responses:
        '200':
          $ref: '#/components/responses/Status'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/SendFailure'
components:
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      description: |
        Optional channel-scoped retry identity. Empty disables deduplication. Reusing
        a nonempty key with the same normalized message, retry limit, and requested
        schedule returns the original ID. Conflicting content returns 409. Keys remain
        effective while their rows are retained. See docs/queue-contract.md for canonical comparison.
      schema:
        type: string
        maxLength: 200
        pattern: '^[!-~]*$'
    MessageID:
      name: id
      in: query
      required: true
      schema:
        type: string
        minLength: 1
    MaxRetries:
      name: max_retries
      in: query
      description: Defaults to zero; any negative integer normalizes to -1 for unlimited retries.
      schema:
        type: integer
        default: 0
    SendAt:
      name: send_at
      in: query
      description: RFC3339 timestamp. Missing or past values queue immediately.
      schema:
        type: string
        format: date-time
  responses:
    Status:
      description: Persisted message status; lookup is not restricted by the route channel
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/MessageStatus'
    NotFound:
      description: No message exists for the ID
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Queued:
      description: Committed to PostgreSQL or matched an identical keyed submission; broker availability is not required and this is not a delivery receipt
      content:
        application/json:
          schema:
            type: object
            required: [id]
            properties:
              id:
                type: string
                format: uuid
    BadRequest:
      description: Invalid binding, query, attachment encoding, idempotency key, or message
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Conflict:
      description: The channel-scoped idempotency key belongs to a different submission
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    SendFailure:
      description: Submission could not be saved; retry with an idempotency key after an uncertain response
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unavailable:
      description: No sender is enabled for the channel, or shutdown has begun
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
...
    HTTPMessage:
      type: object
      required: [url]
      properties:
        url:
          type: string
          description: Absolute HTTP or HTTPS destination without userinfo or fragment; operator allowlisting is enforced at delivery.
        method:
          type: string
          default: POST
        headers:
          type: object
          additionalProperties:
            type: string
        body:
          type: string
    MessageStatus:
      type: object
      required: [id, channel, status, status_text, retry_count, max_retries]
      properties:
        id:
          type: string
          format: uuid
        channel:
          type: string
        status:
          type: integer
          enum: [1, 2, 3, 4, 5]
          description: 1 pending, 2 sending (leased attempt), 3 failed (retry pending), 4 sent (provider accepted), 5 dead (budget or age exhausted).
        status_text:
          type: string
          enum: [pending, sending, failed, sent, dead]
        retry_count:
          type: integer
          description: Attempts after the initial attempt; expired leases consume the attempt budget.
        max_retries:
          type: integer
        next_retry_at:
          description: Initial future schedule or next retry due time; cleared at terminal completion.
          type: string
          format: date-time
        last_error:
          type: string
        sent_at:
          description: Time provider acceptance was persisted, not recipient receipt.
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    Error:
      type: object
      required: [status, message, details]
      properties:
        status:
          type: integer
        message:
          type: string
        details:
          type: [object, 'null']
          additionalProperties: true